--- src/ZendeskApi.Client.Abstractions/Models/TicketType.cs
@@ -1,14 +1,0 @@
-﻿using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-
-namespace ZendeskApi.Client.Models
-{
-    [JsonConverter(typeof(StringEnumConverter))]
-    public enum TicketType
-    {
-        task,
-        incident,
-        problem,
-        question
-    }
-}
