+++ src/ZendeskApi.Client.Abstractions/Models/Tickets/TicketStatus.cs
@@ -1,9 +1,11 @@
 ﻿using System.Runtime.Serialization;
 
-namespace ZendeskApi.Client.Models
+namespace ZendeskApi.Client.Models.Tickets
 {
     public enum TicketStatus
     {
+        [EnumMember(Value = null)]
+        None,
         [EnumMember(Value = "new")]
         New,
         [EnumMember(Value = "open")]
